# docs/openapi-crm-proposal.yaml
openapi: 3.0.3
info:
  title: LM FIT CRM API (proposta)
  version: 0.1.0
  description: >
    Contratos sugeridos para lmfit-api. O painel web (lmfit-web) já compõe dados de cliente
    com pedidos/notas/escalações; estes endpoints consolidam CRM quando implementados.
servers:
  - url: http://127.0.0.1:4000
    description: Dev
...
paths:
...
  /public/payments/{id}:
    get:
      tags: [public-ecommerce]
      summary: Status do pagamento (polling)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, paid, expired, failed, cancelled]
  /public/payments/{id}/dev-confirm:
    post:
      tags: [public-ecommerce]
      summary: Confirma PIX em dev (requer header X-Payment-Dev-Confirm = PAYMENT_DEV_CONFIRM_KEY)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pedido marcado como pago
...
  /reports/sales-today:
    get:
      tags: [reports]
      summary: Vendas do dia (pedidos pagos/atendidos)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string }
                  total: { type: number }
                  orderCount: { type: integer }
                  avgTicket: { type: number }
  /reports/abc:
    get:
      tags: [reports]
      summary: Curva ABC por produto
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date-time }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  range:
                    type: object
                    properties:
                      from: { type: string }
                      to: { type: string }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        productId: { type: string }
                        name: { type: string }
                        sku: { type: string }
                        revenue: { type: number }
                        units: { type: number }
                        cumulativePercent: { type: number }
                        curve: { type: string, enum: [A, B, C] }
...